requirements.txt
streamlit==1.31.0
google-adk==0.0.1
python-dotenv==1.0.0
openai==1.6.1